feat: fsd 구조에 맞춰 컴포넌트 분리
