+++ apps/web/src/entities/auth/ui/enrollment-step-body/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrollmentStepBody';
