--- apps/web/src/entities/auth/ui/enrollment-step-card/index.ts
@@ -1,1 +1,0 @@
-export { EnrollmentStepCard } from './EnrollmentStepCard';
