+++ apps/web/src/entities/auth/ui/index.ts
@@ -1,1 +1,1 @@
-export * from './enrollment-step-card';
+export * from './enrollment-step-body';
