+++ apps/web/src/features/auth/ui/enrollment-step-action-button/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrollmentStepActionButton';
