+++ apps/web/src/features/auth/ui/index.ts
@@ -1,3 +1,4 @@
 export { KakaoLoginButton } from './kakao-login-button';
 export { AppleLoginButton } from './apple-login-button';
 export * from './enrollment-verification-section';
+export * from './enrollment-step-action-button';
