+++ apps/web/src/widgets/auth/ui/enrollment-step-card-header/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrollmentStepCardHeader';
